Fix platform 'allowed groups' race condition on operator upgrade
Commit 80d0c8e8ee8b1a6dd6c06098be8ba06e8f3fe6ac introduced a
subtle bug where, upon upgrading the operator to version 0.9.0
using standard Azimuth tooling, the `/platform-users` group is
inadvertently dropped from the allowed groups list for all
platforms deployed prior to the upgrade. This causes HTTP 403
errors in the browser when a user attempts to access any Zenith
service for the previously deployed platform.

The root cause of this bug is as follows:

During an Azimuth upgrade, when the new identity operator chart
gets installed it restarts the identity operator container and
triggers the ‘on resume’ handler for both the Realm and Platform
CRDs. This means that both realms and platforms get reconciled at
the same time so the Platform resources get reconciled before the
operator has added the status.platform_users_group field to the
Realm and this causes /platform-users to be removed from the
allowed groups.

The fix proposed here is to fall back to pre-0.9.0 behaviour if
the realm.status.platform_users_group field is empty. This
preserves backwards compatibility with existing platforms but
could also be removed in a future 0.10.0 release assuming that
by that point all realms have been reconciled at least once using
the 0.9.0 behaviour in which the status field is populated.